# server/dist/swagger.json
{
	"basePath": "/api",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"ILoginResponse": {
			"properties": {
				"authToken": {
					"type": "string"
				},
				"username": {
					"type": "string"
				},
				"id": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"authToken",
				"username",
				"id"
			],
			"type": "object"
		},
		"IErrorResponse": {
			"properties": {
				"thrown": {
					"type": "boolean",
					"x-nullable": true
				},
				"error": {
					"type": "object",
					"x-nullable": true
				},
				"message": {
					"type": "string",
					"x-nullable": true
				}
			},
			"type": "object"
		},
		"ILoginParams": {
			"properties": {
				"username": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}
			},
			"required": [
				"username",
				"password"
			],
			"type": "object"
		},
		"IUserResponse": {
			"properties": {
				"username": {
					"type": "string"
				},
				"id": {
					"type": "number",
					"format": "double"
				},
				"nsfwtags": {
					"type": "string"
				}
			},
			"required": [
				"username",
				"id",
				"nsfwtags"
			],
			"type": "object"
		},
		"INewUserParams": {
			"properties": {
				"username": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}
			},
			"required": [
				"username",
				"password"
			],
			"type": "object"
		},
		"PartialIUser": {
			"description": "Make all properties in T optional",
			"properties": {},
			"type": "object"
		},
		"ITag": {
			"properties": {
				"id": {
					"type": "number",
					"format": "double",
					"x-nullable": true
				},
				"value": {
					"type": "string"
				},
				"pictures": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/IPicture"
					},
					"x-nullable": true
				}
			},
			"required": [
				"value"
			],
			"type": "object"
		},
		"IPicture": {
			"properties": {
				"id": {
					"type": "number",
					"format": "double",
					"x-nullable": true
				},
				"creationDate": {
					"type": "string",
					"format": "date-time"
				},
				"tags": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/ITag"
					},
					"x-nullable": true
				},
				"votes": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/IVote"
					}
				},
				"data": {
					"type": "string",
					"format": "byte"
				},
				"nsfwTags": {
					"type": "string"
				},
				"location": {
					"type": "object"
				}
			},
			"required": [
				"creationDate",
				"votes",
				"data",
				"nsfwTags",
				"location"
			],
			"type": "object"
		},
		"IUser": {
			"properties": {
				"username": {
					"type": "string"
				},
				"password": {
					"type": "string",
					"x-nullable": true
				},
				"nsfwtags": {
					"type": "string"
				},
				"id": {
					"type": "number",
					"format": "double",
					"x-nullable": true
				}
			},
			"required": [
				"username",
				"nsfwtags"
			],
			"type": "object"
		},
		"IVote": {
			"properties": {
				"id": {
					"type": "number",
					"format": "double",
					"x-nullable": true
				},
				"up": {
					"type": "boolean"
				},
				"picture": {
					"$ref": "#/definitions/IPicture"
				},
				"value": {
					"type": "number",
					"format": "double"
				},
				"pictureId": {
					"type": "number",
					"format": "double"
				},
				"user": {
					"$ref": "#/definitions/IUser"
				},
				"userId": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"up",
				"picture",
				"value",
				"pictureId",
				"user",
				"userId"
			],
			"type": "object"
		},
		"IPost": {
			"properties": {
				"id": {
					"type": "string"
				},
				"pictureID": {
					"type": "string"
				},
				"nsfwTags": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"tags": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"votes": {
					"type": "number",
					"format": "double"
				},
				"coordinates": {
					"type": "object"
				}
			},
			"required": [
				"id",
				"pictureID",
				"nsfwTags",
				"tags",
				"votes",
				"coordinates"
			],
			"type": "object"
		},
		"IFeedResponse": {
			"properties": {
				"pictures": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/IPost"
					}
				},
				"page": {
					"type": "number",
					"format": "double"
				},
				"pageSize": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"pictures",
				"page",
				"pageSize"
			],
			"type": "object"
		},
		"ILocationInfo": {
			"properties": {
				"lat": {
					"type": "number",
					"format": "double"
				},
				"lon": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"lat",
				"lon"
			],
			"type": "object"
		},
		"IFeedRequestBody": {
			"properties": {
				"page": {
					"type": "number",
					"format": "double"
				},
				"pageSize": {
					"type": "number",
					"format": "double"
				},
				"location": {
					"$ref": "#/definitions/ILocationInfo"
				},
				"tags": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"x-nullable": true
				}
			},
			"required": [
				"page",
				"pageSize",
				"location"
			],
			"type": "object"
		},
		"IImageParams": {
			"properties": {
				"tags": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"coords": {
					"type": "string"
				}
			},
			"required": [
				"tags",
				"coords"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "PikPak-Server",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		}
	},
	"paths": {
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/ILoginResponse"
						}
					},
					"default": {
						"description": "Error Occurred",
						"schema": {
							"$ref": "#/definitions/IErrorResponse"
						}
					}
				},
				"description": "Login user",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"description": "Parameters to login",
						"in": "body",
						"name": "loginParams",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ILoginParams"
						}
					}
				]
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "RegisterUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IUserResponse"
						}
					},
					"default": {
						"description": "Error occurred",
						"schema": {
							"$ref": "#/definitions/IErrorResponse"
						}
					}
				},
				"description": "Register new user",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"description": "Parameters to register new user",
						"in": "body",
						"name": "requestBody",
						"required": true,
						"schema": {
							"$ref": "#/definitions/INewUserParams"
						}
					}
				]
			}
		},
		"/auth": {
			"get": {
				"operationId": "CurrentUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IUserResponse"
						}
					},
					"default": {
						"description": "Error occurred",
						"schema": {
							"$ref": "#/definitions/IErrorResponse"
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"JWT": []
					}
				],
				"parameters": []
			}
		},
		"/auth/update": {
			"post": {
				"operationId": "UpdateUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IUserResponse"
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "user",
						"required": true,
						"schema": {
							"$ref": "#/definitions/PartialIUser"
						}
					}
				]
			}
		},
		"/feed/tags": {
			"get": {
				"operationId": "GetTags",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/ITag"
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/feed": {
			"post": {
				"operationId": "Search",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IFeedResponse"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IFeedRequestBody"
						}
					}
				]
			}
		},
		"/version": {
			"get": {
				"operationId": "Version",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/image/imageFile/{fileId}.jpeg": {
			"get": {
				"operationId": "GetAttachment",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "fileId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/image/vote": {
			"get": {
				"operationId": "Vote",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IPost"
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "up",
						"required": true,
						"type": "boolean"
					}
				]
			}
		},
		"/image/getVoteInfo": {
			"get": {
				"operationId": "GetVotes",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/IVote"
							}
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				],
				"parameters": []
			}
		},
		"/image/uploadFile": {
			"post": {
				"operationId": "UploadFile",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IPost"
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				],
				"parameters": [
					{
						"in": "formData",
						"name": "imageFile",
						"required": true,
						"type": "file"
					},
					{
						"in": "formData",
						"name": "tags",
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					}
				],
				"consumes": [
					"multipart/form-data"
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"JWT": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	}
}